requirements.txt
altgraph==0.17
appdirs==1.4.4
cycler==0.10.0
kiwisolver==1.3.1
macholib==1.14
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyopencl==2021.1.6
pyparsing==2.4.7
PySide2==5.15.2
python-dateutil==2.8.1
pytools==2021.2.6
shiboken2==5.15.2
six==1.15.0
